fix: rename terminal client to sls
